# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GardenQuest Backend API
  version: 1.0.0
  description: |
    Especificacao OpenAPI das rotas HTTP atualmente expostas pelos entrypoints
    `backend/server.js` (modo legado) e `backend/api-server.js` (modo API + Worker).

    Observacoes importantes:
    - Autenticacao principal via cookie `auth_token` (JWT + sessao revogavel).
    - Parte das rotas ainda usa erro legado `{ "error": "..." }`.
    - Rotas de sistema (`/api/v1/system/*`) usam payload padronizado com correlation-id.
    - As rotas `/api/v1/games/garden-quest/*` sao alias das rotas `/api/v1/ai-game/*`.
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
...
paths:
...
  /auth/google:
    get:
      tags: [auth]
      summary: Inicia login Google OAuth
      operationId: startGoogleOAuth
      parameters:
        - in: query
          name: redirect
          required: false
          schema:
            type: string
          description: "Path frontend de retorno (ex.: `/hub.html`)."
      responses:
        '302':
          description: Redireciona para Google OAuth.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Cookie `oauth_state` com nonce anti-CSRF.
            Location:
              schema:
                type: string
  /auth/callback:
    get:
      tags: [auth]
      summary: Callback OAuth
      operationId: completeGoogleOAuth
      parameters:
        - in: query
          name: code
          required: false
          schema:
            type: string
        - in: query
          name: state
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redireciona para frontend com sucesso/erro de auth.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Define/limpa `auth_token` e limpa `oauth_state`.
            Location:
              schema:
                type: string
  /auth/me:
    get:
      tags: [auth]
      summary: Retorna perfil autenticado
      operationId: getAuthMe
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Usuario autenticado.
          headers:
            x-correlation-id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /auth/sessions:
    get:
      tags: [auth]
      summary: Lista sessoes ativas do usuario
      operationId: listAuthSessions
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Lista de sessoes.
          headers:
            x-correlation-id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /auth/sessions/{sessionId}/revoke:
    post:
      tags: [auth]
      summary: Revoga uma sessao do proprio usuario
      operationId: revokeOwnAuthSession
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sessao revogada (ou estado final retornado).
          headers:
            x-correlation-id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionRevokeResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /auth/logout-all:
    post:
      tags: [auth]
      summary: Revoga todas as sessoes do usuario
      operationId: logoutAllSessions
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout global concluido.
          headers:
            x-correlation-id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutAllResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  headers:
    CorrelationId:
      description: Correlation id propagado pelo middleware `request-context`.
      schema:
        type: string
  responses:
...
    BadRequestResponse:
      description: Requisicao invalida.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedResponse:
      description: Sessao ausente, expirada ou revogada.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundResponse:
      description: Recurso nao encontrado.
      headers:
        x-correlation-id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    AuthMeResponse:
      type: object
      required: [id, name, email, isAdmin]
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        picture:
          type: string
          nullable: true
        isAdmin:
          type: boolean
    SessionRecord:
      type: object
      properties:
        id:
          type: string
        userEmail:
          type: string
          nullable: true
        userName:
          type: string
          nullable: true
        ip:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
        issuedAt:
          type: string
          nullable: true
        expiresAt:
          type: string
          nullable: true
        lastSeenAt:
          type: string
          nullable: true
        revokedAt:
          type: string
          nullable: true
        revokeReason:
          type: string
          nullable: true
        isCurrent:
          type: boolean
    AuthSessionsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionRecord'
    AuthSessionRevokeResponse:
      type: object
      required: [ok, revoked, session]
      properties:
        ok:
          type: boolean
        revoked:
          type: boolean
        session:
          $ref: '#/components/schemas/SessionRecord'
    LogoutAllResponse:
      type: object
      required: [ok, revokedCount]
      properties:
        ok:
          type: boolean
        revokedCount:
          type: integer
...
    LegacyError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CorrelatedError:
      type: object
      required: [error, code, errorId, correlationId]
      properties:
        error:
          type: string
        code:
          type: string
          example: internal_error
        errorId:
          type: string
          nullable: true
        correlationId:
          type: string
          nullable: true
    ErrorResponse:
      oneOf:
        - $ref: '#/components/schemas/LegacyError'
        - $ref: '#/components/schemas/CorrelatedError'
...